+++ src/constants/customHeaderKeys.ts
@@ -1,0 +1,2 @@
+export const X_HEADER_TITLE = 'x-header-title';
+export const X_PATH_NAME = 'x-path-name';
